[Refactoring] Winning.java - Winning도메인 리팩토링
