+++ src/main/java/lotto/service/Statistics.java
@@ -1,0 +1,4 @@
+package lotto.service;
+
+public class Statistics {
+}
