+++ src/test/java/lotto/purchaseTest.java
@@ -1,6 +1,7 @@
-package lotto.service;
+package lotto;
 
 import lotto.domain.Lotto;
+import lotto.service.Purchase;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -10,7 +11,7 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
-class purchaseServiceTest {
+class purchaseTest {
 
     @Nested
     @DisplayName("구매금액 입력오류시, 생성 예외발생")
@@ -22,27 +23,27 @@ class purchaseServiceTest {
         @Test
         void test1() {
             purchaseAmount = "78000";
-            assertDoesNotThrow(() -> new purchaseService(purchaseAmount));
+            assertDoesNotThrow(() -> new Purchase(purchaseAmount));
 
             purchaseAmount = "1000";
-            assertDoesNotThrow(() -> new purchaseService(purchaseAmount));
+            assertDoesNotThrow(() -> new Purchase(purchaseAmount));
         }
 
         @DisplayName("구매금액이 문자를 포함할 경우 예외를 발생")
         @Test
         void test2() {
             purchaseAmount = "10O00";
-            assertThrows(IllegalArgumentException.class, () -> new purchaseService(purchaseAmount));
+            assertThrows(IllegalArgumentException.class, () -> new Purchase(purchaseAmount));
         }
 
         @DisplayName("구매금액이 1000원 단위가 아닐 경우 예외를 발생")
         @Test
         void test3() {
             purchaseAmount = "500";
-            assertThrows(IllegalArgumentException.class, () -> new purchaseService(purchaseAmount));
+            assertThrows(IllegalArgumentException.class, () -> new Purchase(purchaseAmount));
 
             purchaseAmount = "34200";
-            assertThrows(IllegalArgumentException.class, () -> new purchaseService(purchaseAmount));
+            assertThrows(IllegalArgumentException.class, () -> new Purchase(purchaseAmount));
         }
 
     }
@@ -50,8 +51,8 @@ class purchaseServiceTest {
     @DisplayName("13000원 구매시, 13개의 로또가 발행되어야 한다.")
     @Test
     void issueLotteries() {
-        purchaseService purchaseService = new purchaseService("13000");
-        List<Lotto> lotteries = purchaseService.issueLotteries();
+        Purchase Purchase = new Purchase("13000");
+        List<Lotto> lotteries = Purchase.issueLotteries();
         assertThat(13).isEqualTo(lotteries.size());
     }
